+++ websitebuild-20250429/ChangeLog.md
@@ -1,3 +1,9 @@
+2026-04-20 Version: 2.20.1
+- Update API GetAppInstanceForPartner: add response parameters Body.Module.AppServiceList.$.Profile.UpgradeStatus.
+- Update API GetAppInstanceForPartner: add response parameters Body.Module.Profile.UpgradeStatus.
+- Update API GetAppInstanceForPartner: add response parameters Body.Module.RelatedInstanceList.$.Profile.UpgradeStatus.
+
+
 2026-04-04 Version: 2.20.0
 - Support API CheckUserResourceMeasure.
 - Support API CreateAppAssistantAgent.